requirements.txt
Django==3.2.8
djangorestframework==3.12.4
drf_yasg==1.20.0
requests==2.26.0
